catvsdog_model_api/requirements.txt
uvicorn==0.17.6
fastapi==0.99.1
pydantic==1.10.10
requests==2.23.0
jinja2==3.1.2
python-multipart==0.0.6
pillow==9.5.0

# repo maintenance tooling
#black>=22.0.0,<23.0.0
#flake8>=3.9.0,<3.10.0
#mypy==0.812
#isort==5.8.0

# After uploading your package in PyPi use the given below
#catvsdog-model==0.0.1

# If locally copied whl file inside catvsdog_model_api then use given below
catvsdog_model-0.0.1-py3-none-any.whl